@@ -22,72 +18,94 @@
     "react",
     "component",
     "element-queries",
+    "element-query",
     "size",
     "dimensions"
   ],
-  "author": "procoder17 <procoder17@outlook.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/procoder17/react-sizeme/issues"
-  },
   "homepage": "https://github.com/procoder17/react-sizeme#readme",
-  "devDependencies": {
-    "babel-cli": "6.14.0",
-    "babel-core": "6.14.0",
-    "babel-eslint": "6.1.2",
-    "babel-loader": "6.2.5",
-    "babel-plugin-lodash": "3.2.8",
-    "babel-plugin-rewire": "1.0.0-rc-2",
-    "babel-preset-es2015": "6.14.0",
-    "babel-preset-react": "6.11.1",
-    "babel-preset-stage-1": "6.13.0",
-    "babel-register": "6.14.0",
-    "chai": "3.5.0",
-    "codecov.io": "0.1.6",
-    "commitizen": "2.8.6",
-    "compression": "1.6.2",
-    "cz-conventional-changelog": "1.2.0",
-    "enzyme": "2.4.1",
-    "eslint": "3.5.0",
-    "eslint-config-airbnb": "11.1.0",
-    "eslint-loader": "1.5.0",
-    "eslint-plugin-import": "1.14.0",
-    "eslint-plugin-jsx-a11y": "2.2.2",
-    "eslint-plugin-mocha": "4.5.1",
-    "eslint-plugin-react": "6.2.0",
-    "express": "4.14.0",
-    "ghooks": "1.3.2",
-    "isparta": "4.0.0",
-    "jsdom": "9.5.0",
-    "lodash-webpack-plugin": "0.10.0",
-    "mocha": "3.0.2",
-    "path": "0.12.7",
-    "react": "15.3.1",
-    "react-addons-test-utils": "15.3.1",
-    "react-dom": "15.3.1",
-    "semantic-release": "4.3.5",
-    "sinon": "1.17.5",
-    "sinon-chai": "2.8.0",
-    "stats-webpack-plugin": "0.4.2",
-    "webpack": "1.13.2",
-    "webpack-dev-middleware": "1.7.0",
-    "webpack-hot-middleware": "2.12.2"
+  "author": "procoder17 <procoder17@outlook.com>",
+  "keywords": [
+    "library"
+  ],
+  "scripts": {
+    "build": "babel-node ./tools/scripts/build.js",
+    "check": "yarn run lint && yarn run test",
+    "clean": "rimraf ./commonjs && rimraf ./umd && rimraf ./coverage && rimraf ./umd",
+    "example:web": "echo 'Make sure to `cd example/web && yarn install`' && cd example/web && yarn run start",
+    "lint": "eslint src",
+    "precommit": "lint-staged && npm run test",
+    "prepublish": "yarn run build",
+    "test": "jest",
+    "test:coverage": "yarn run test -- --coverage",
+    "test:coverage:deploy": "yarn run test:coverage && codecov"
   },
-  "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0",
-    "react-dom": "^0.14.0 || ^15.0.0"
+  "jest": {
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx}"
+    ],
+    "testPathIgnorePatterns": [
+      "<rootDir>/(commonjs|coverage|flow-typed|node_modules|tools|umd)/"
+    ]
   },
-  "config": {
-    "ghooks": {
-      "pre-commit": "npm run test"
-    },
-    "commitizen": {
-      "path": "node_modules/cz-conventional-changelog"
-    }
+  "lint-staged": {
+    "*.js": [
+      "prettier-eslint --write",
+      "git add"
+    ]
   },
   "dependencies": {
-    "element-resize-detector": "1.1.8",
-    "invariant": "^2.2.0",
-    "lodash": "4.15.0"
+    "element-resize-detector": "^1.1.11",
+    "invariant": "^2.2.2",
+    "lodash": "^4.17.4"
+  },
+  "peerDependencies": {
+    "react": "^0.14.0 || ^15.0.0-0",
+    "react-dom": "^0.14.0 || ^15.0.0-0"
+  },
+  "devDependencies": {
+    "app-root-dir": "1.0.2",
+    "babel-cli": "6.24.0",
+    "babel-core": "6.24.0",
+    "babel-eslint": "7.2.1",
+    "babel-jest": "19.0.0",
+    "babel-loader": "6.4.1",
+    "babel-plugin-lodash": "3.2.11",
+    "babel-plugin-rewire": "1.0.0",
+    "babel-plugin-transform-class-properties": "6.23.0",
+    "babel-polyfill": "6.23.0",
+    "babel-preset-env": "1.2.2",
+    "babel-preset-latest": "6.24.0",
+    "babel-preset-react": "6.23.0",
+    "babel-preset-stage-3": "6.22.0",
+    "babel-register": "6.24.0",
+    "codecov": "2.1.0",
+    "cross-env": "3.2.4",
+    "enzyme": "2.8.0",
+    "enzyme-to-json": "1.5.0",
+    "eslint": "3.18.0",
+    "eslint-config-airbnb": "14.1.0",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-jsx-a11y": "4.0.0",
+    "eslint-plugin-react": "6.10.3",
+    "gzip-size": "3.0.0",
+    "husky": "0.13.3",
+    "in-publish": "2.0.0",
+    "jest": "19.0.2",
+    "lint-staged": "3.4.0",
+    "lodash-webpack-plugin": "0.11.2",
+    "prettier": "0.22.0",
+    "prettier-eslint": "4.3.2",
+    "prettier-eslint-cli": "3.1.3",
+    "pretty-bytes": "4.0.2",
+    "ramda": "0.23.0",
+    "react": "15.4.2",
+    "react-addons-test-utils": "15.4.2",
+    "react-dom": "15.4.2",
+    "readline-sync": "1.4.7",
+    "rimraf": "2.6.1",
+    "sinon": "2.1.0",
+    "webpack": "2.3.2",
+    "webpack-dev-middleware": "1.10.1",
+    "webpack-hot-middleware": "2.17.1"
   }
 }
