@@ -1,18 +1,11 @@
-import React, { PropTypes } from 'react';
-
-let sizeMe;
-
-if (process.env.NODE_ENV === 'development') {
-  sizeMe = require('../../src/sizeMe.js').default;
-} else {
-  sizeMe = require('react-sizeme').default;
-}
+import React, { PropTypes } from 'react'
+import sizeMe from 'react-sizeme'
 
 const rootStyle = {
   fontWeight: 'bold',
   position: 'relative',
   textAlign: 'center',
-};
+}
 
 const spanStyle = {
   position: 'absolute',
@@ -20,18 +13,20 @@ const spanStyle = {
   top: '50%',
   left: '50%',
   transform: 'translateX(-50%) translateY(-50%)',
-};
+}
 
 function MyComponent({ children, size: { width, height }, style }) {
   return (
     <div style={Object.assign({}, rootStyle, style)}>
       <span style={spanStyle}>
         {Math.round(width)}x{Math.round(height)}<br />
-        <span style={{ fontWeight: 'normal', fontStyle: 'italic' }}>(rounded)</span>
+        <span style={{ fontWeight: 'normal', fontStyle: 'italic' }}>
+          (rounded)
+        </span>
       </span>
       {children}
     </div>
-  );
+  )
 }
 
 MyComponent.propTypes = {
@@ -41,10 +36,8 @@ MyComponent.propTypes = {
     height: PropTypes.number.isRequired,
   }),
   style: PropTypes.object,
-};
+}
 
 export default sizeMe({
   monitorHeight: true,
-  refreshRate: 2500,
-  refreshMode: 'debounce'
-})(MyComponent);
+})(MyComponent)
